+++ CHANGELOG.md
@@ -9,6 +9,7 @@ principles.
 -   Added: you can now find the current Chief version on the bottom of the sidebar navigation view
 -   Fixed: field-in-window flexbox items are sometimes overflowing the container (e.g. repeat field window view)
 -   Fixed: RadioFilter and CheckboxFilter field labels were not correctly styled
+-   Fixed: Checkbox form-input-toggle doesn't shrink anymore in flexbox with long label
 
 ## 0.7.22 - 2023-03-13